backend/app/utils/sign_util.py
import logging
from flask import current_app

# 设置日志记录器
logger = logging.getLogger(__name__)

# 假设我们使用某个CLOB客户端库，这里需要根据实际情况导入
# 例如：from clob_client import ClobClient
# 由于具体的CLOB客户端库可能不同，这里使用通用的方式实现
try:
    from clob_client import ClobClient
except ImportError:
    logger.warning("CLOB客户端库未安装，可能需要安装相关依赖")
    ClobClient = None


def init_clob_client(private_key, wallet_address):
    """初始化CLOB客户端
    
    Args:
        private_key: 钱包私钥
        wallet_address: 钱包地址
    
    Returns:
        ClobClient实例
    
    Raises:
        ImportError: 如果CLOB客户端库未安装
        Exception: 如果客户端初始化失败
    """
    if not ClobClient:
        raise ImportError("CLOB客户端库未安装，请先安装相关依赖")
    
    try:
        # 使用应用配置中的API URL，如果没有则使用默认URL
        base_url = current_app.config.get('CLOB_API_URL', 'https://api.clob.com')
        
        logger.info(f"初始化CLOB客户端，API URL: {base_url}")
        
        # 创建并返回CLOB客户端实例
        # 这里的参数可能需要根据实际的ClobClient构造函数进行调整
        client = ClobClient(
            private_key=private_key,
            address=wallet_address,
            base_url=base_url
        )
        
        logger.info("CLOB客户端初始化成功")
        return client
    except Exception as e:
        logger.error(f"CLOB客户端初始化失败: {str(e)}")
        raise Exception(f"CLOB客户端初始化失败: {str(e)}") from e
