backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.29
Celery==5.3.6
PyJWT==2.8.0
passlib==1.7.4
flask-cors==4.0.1
flask-jwt-extended==4.6.0
requests==2.31.0
python-dotenv==1.0.1
pymysql==1.1.0
redis==5.0.3
celery[redis]==5.3.6
